Revert "DB updated"
