+++ project/package.json
@@ -12,7 +12,6 @@
     "antd": "^5.10.3",
     "antd-img-crop": "^4.20.0",
     "axios": "^1.6.5",
-    "bootstrap": "^5.3.3",
     "font-awesome": "^4.7.0",
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.1",
