+++ project/public/index.html
@@ -26,7 +26,7 @@
     -->
     <title>Sports Management System</title>
   </head>
-  <body style="height: 100vh; overflow-y: auto;">
+  <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <!--
