+++ project/src/Components/Coach/CoachCreateTeam/CoachCreateTeam.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import './CoachCreateTeam.css'
+import CoachSidebar from '../CoachSidebar/CoachSidebar'
+
+
+const CoachCreateTeam = () => {
+  return (
+    <div>
+      <CoachSidebar>
+        
+      </CoachSidebar>
+    </div>
+  )
+}
+
+export default CoachCreateTeam
