+++ project/src/Components/Coach/CoachEditTeam/CoachEditTeam.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import "./CoachEditTeam.css";
+import CoachSlider from "../CoachSidebar/CoachSidebar";
+
+const CoachEditTeam = () => {
+  return (
+    <div>
+      <CoachSlider>
+        
+      </CoachSlider>
+    </div>
+  );
+};
+
+export default CoachEditTeam;
