+++ project/src/Components/Coach/CoachSidebar/CoachSidebar.jsx
@@ -211,20 +211,16 @@ const currentUserData = async () => {
                   <span className="nav-text">My Availability</span>
                 </NavLink>
               </Menu.Item>
-
-              
-               <Menu.Item key="/coach-create-team" icon={<EditOutlined />}>
-                <NavLink to="/create-team">
+              {/* <Menu.Item key="/coach-create-team" icon={<EditOutlined />}>
+                <NavLink to="/coach-create-team">
                   <span className="nav-text">Create Team</span>
                 </NavLink>
               </Menu.Item>
               <Menu.Item key="/coach-edit-team" icon={<FormOutlined />}>
-                <NavLink to="/edit-team">
+                <NavLink to="/coach-edit-team">
                   <span className="nav-text">Edit Team</span>
                 </NavLink>
-              </Menu.Item>
-
-
+              </Menu.Item> */}
               <Menu.Item
                 key="/coach-review-players"
                 icon={<CalendarOutlined />}
