+++ project/src/Components/EventOrganizer/EOBracket/EOBracket.jsx
@@ -51,13 +51,6 @@ const Bracket = () => {
 
   return (
     <EOSizeBar>
-
-
-
-
-
-
-      {/*
       <div className="outer-container">
         <div className="inputContainer">
           <label>
@@ -112,8 +105,6 @@ const Bracket = () => {
           </div>
         )}
       </div>
-
-                      */}
     </EOSizeBar>
   );
 };
