+++ project/src/Pages/Coach/UpdateTeam.jsx
@@ -3,7 +3,6 @@
 import axios from 'axios';
 import React, { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import CoachSidebar from '../CoachSidebar/CoachSidebar';
 
 const UpdateTeam = () => {
     const navigate = useNavigate()
@@ -27,7 +26,7 @@ const UpdateTeam = () => {
     const [allplayers, setAllplayers] = useState([]);
 
     useEffect(() => {
-        axios.get(`http://localhost:8080/api/v1/coach/players?match_id=${match_id}&coach_id=${coach_id}`)
+        axios.get(`http://localhost:8080/coach/players?match_id=${match_id}&coach_id=${coach_id}`)
             .then(res => {
                 console.log(res.data);
                 setAllplayers(res.data);
@@ -60,7 +59,7 @@ const UpdateTeam = () => {
     const handleSubmit = async () => {
         try {
             console.log('edited team data to be sent : ', teamData)
-            const res = await axios.post(`http://localhost:8080/api/v1/coach/update-team?team_id=${team_id}`, { teamData: teamData });
+            const res = await axios.post(`http://localhost:8080/coach/update-team?team_id=${team_id}`, {teamData: teamData});
             if (res.data.success) {
                 console.log('Team data updated successfully: ', res.data.team);
             } else {
@@ -80,47 +79,45 @@ const UpdateTeam = () => {
     };
 
     return (
-        <CoachSidebar>
-            <div style={{ height: '100%', overflow: 'auto' }}>
-                <div style={{ margin: 20, padding: 20, textAlign: 'center' }}>Update Team <br />
-                    <form>
-                        Team Name : <input type="text" name='teamName' value={teamData.teamName} onChange={handleChange} />
-                        Team No : <input type="text" name='teamNo' value={teamData.teamNo} onChange={handleChange} />
-                    </form>
-                </div>
-
-                <div style={{ margin: 10, padding: 10 }}>
-                    <table style={{ border: 'solid 1px grey', width: '100%' }}>
-                        <thead>
-                            <tr style={{ backgroundColor: 'darkblue', color: 'white', textAlign:'center' }}>
-                                <th>Player No</th><th>Player Name</th><th>Location</th><th>Action</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {allplayers.map((player, index) => (
-                                (player.Status === 'available' || player.matches.includes(match_id)) && (<tr  style={{fontWeight:'bold'}} key={index}>
-                                    <td style={{ textAlign: 'center' }} >{player.PlayerNo}</td>
-                                    <td style={{ textAlign: 'center' }}>{player.playerName}</td>
-                                    <td style={{ textAlign: 'center' }}>{player.District}</td>
-                                    <td style={{ textAlign: 'center' }}>
-                                        <button className='btn btn-success' onClick={() => handleAdd(player._id)}>
-                                            {teamData.players.includes(player._id) ? 'Remove' : 'Add'}
-                                        </button>
-                                    </td>
-                                </tr>)
-                            ))}
-                        </tbody>
-                    </table>
-                </div>
-
-                {/* Display submit button only if there are selected players */}
-                {(
-                    <div style={{ textAlign: 'center', marginTop: '20px' }}>
-                        <button className='btn btn-success' onClick={handleSubmit}>Submit</button>
-                    </div>
-                )}
+        <div>
+            <div style={{ margin: 20, padding: 20, textAlign: 'center' }}>Update Team <br />
+                <form>
+                    Team Name : <input type="text" name='teamName' value={teamData.teamName} onChange={handleChange} />
+                    Team No : <input type="text" name='teamNo' value={teamData.teamNo} onChange={handleChange} />
+                </form>
             </div>
-        </CoachSidebar>
+
+            <div style={{ margin: 10, padding: 10 }}>
+                <table style={{ border: 'solid 1px grey', width: '100%' }}>
+                    <thead>
+                        <tr style={{ backgroundColor: 'darkblue', color: 'white' }}>
+                            <th>Player No</th><th>Player Name</th><th>Location</th><th>Action</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        {allplayers.map((player, index) => (
+                            (player.Status === 'available' || player.matches.includes(match_id))&&(<tr key={index}>
+                                <td style={{ textAlign: 'center' }} >{player.PlayerNo}</td>
+                                <td style={{ textAlign: 'center' }}>{player.FirstName}</td>
+                                <td style={{ textAlign: 'center' }}>{player.District}</td>
+                                <td style={{ textAlign: 'center' }}>
+                                    <button onClick={() => handleAdd(player._id)}>
+                                        {teamData.players.includes(player._id) ? 'Remove' : 'Add'}
+                                    </button>
+                                </td>
+                            </tr>)
+                        ))}
+                    </tbody>
+                </table>
+            </div>
+
+            {/* Display submit button only if there are selected players */}
+            {(
+                <div style={{ textAlign: 'center', marginTop: '20px' }}>
+                    <button onClick={handleSubmit}>Submit</button>
+                </div>
+            )}
+        </div>
     );
 };
