+++ server/controllers/PlayerAvailabilityController/PlayerAvailabilityController.js
@@ -1,5 +1,4 @@
 const PlayerAvailabilityModel = require("../../models/PlayerAvailabilityModel/PlayerAvailabilityModel");
-const PlayerProfileModel = require('../../models/PlayerProfileModel/PlayerProfileModel')
 
 const PlayerAvailabilityController = async(req,res)=>{
    //   console.log(req.body);
@@ -21,8 +20,6 @@ const PlayerAvailabilityController = async(req,res)=>{
 
                    await setAvailability.save();
 
-                   await PlayerProfileModel.findOneAndUpdate({PlayerNo:playerId}, {$addToSet:{Status:availability}},{ new: true, overwrite: true });
-
                    res.status(200).send({
                            success: true,
                            message: "Availability added successfully",
@@ -34,9 +31,6 @@ const PlayerAvailabilityController = async(req,res)=>{
            else {
                    const setAvailability = await PlayerAvailabilityModel.findOneAndUpdate({ eventId: eventId }, { availability: availability }, { new: true })
                    await setAvailability.save();
-
-                   await PlayerProfileModel.findOneAndUpdate({PlayerNo:playerId}, {$addToSet:{Status:availability}}, { new: true, overwrite: true });
-
 
                    res.status(200).send({
                            success: true,
