+++ server/models/CoachProfileModel/CoachProfileModel.js
@@ -2,26 +2,25 @@ const mongoose = require('mongoose');
 
 const coachProfileSchema = new mongoose.Schema({
 
-   coachNo: String,
-   teams: [{
-     type: mongoose.Schema.Types.ObjectId,
-     ref: 'Team'
-   }],
+     coachId:{
+       type:String,
+//        required:["coach name is Required",true]
+     },
      coachName:{
         type:String,
-        //required:["coach name is Required",true]
+        required:["coach name is Required",true]
      },
      coachEmail:{
         type:String,
-        //required:["coach email is Required",true]
+        required:["coach email is Required",true]
      },
      coachDateOfBirth:{
         type:Object,
-        //required:["coach Date Of Birth is Required",true]
+        required:["coach Date Of Birth is Required",true]
      },
      coachAge:{
         type:Number,
-        //required:["coach age is Required",true]
+        required:["coach age is Required",true]
      },
 
 })
