+++ server/models/teamModel.js
@@ -7,15 +7,15 @@ const teamSchema = new mongoose.Schema({
 
   coach_id: {
     type: mongoose.Schema.Types.ObjectId,
-    ref: 'coachProfile'
+    ref: 'Coach'
   },
   players: [{
     type: mongoose.Schema.Types.ObjectId,
-    ref: 'playerProfile'
+    ref: 'Player'
   }],
   match_id: {
     type: mongoose.Schema.Types.ObjectId,
-    ref: 'ceateEvent'
+    ref: 'Match'
   }
 });
